--- a/modin/pandas/groupby.py
+++ b/modin/pandas/groupby.py
@@ -38,7 +38,7 @@ from modin.core.storage_formats.base.query_compiler import BaseQueryCompiler
 from modin.core.dataframe.algebra.default2pandas.groupby import GroupBy
 from modin.config import IsExperimental
 from .series import Series
-from .utils import is_label
+from .utils import is_label, NumericOnly
 
 
 @_inherit_docstrings(pandas.core.groupby.DataFrameGroupBy)
@@ -124,7 +124,7 @@ class DataFrameGroupBy(ClassLogger):
             type(self._query_compiler).groupby_skew,
             agg_args=args,
             agg_kwargs=kwargs,
-            numeric_only=True,
+            numeric_only=NumericOnly.TRUE_EXCL_NUMERIC_CATEGORIES,
         )
 
     def ffill(self, limit=None):
@@ -1015,7 +1015,7 @@ class DataFrameGroupBy(ClassLogger):
     def _wrap_aggregation(
         self,
         qc_method,
-        numeric_only=None,
+        numeric_only=NumericOnly.AUTO,
         agg_args=None,
         agg_kwargs=None,
         **kwargs,
@@ -1027,12 +1027,8 @@ class DataFrameGroupBy(ClassLogger):
         ----------
         qc_method : callable
             The query compiler method to call.
-        numeric_only : {None, True, False}, default: None
-            Specifies whether to aggregate non numeric columns:
-                - True: include only numeric columns (including categories that holds a numeric dtype)
-                - False: include all columns
-                - None: infer the parameter, ``False`` if there are no numeric types in the frame,
-                  ``True`` otherwise.
+        numeric_only : NumericOnly or its values, default: NumericOnly.AUTO
+            Specifies whether to aggregate non numeric columns.
         agg_args : list-like, optional
             Positional arguments to pass to the aggregation function.
         agg_kwargs : dict-like, optional
@@ -1045,18 +1041,21 @@ class DataFrameGroupBy(ClassLogger):
         DataFrame or Series
             Returns the same type as `self._df`.
         """
+        if not isinstance(numeric_only, NumericOnly):
+            numeric_only = NumericOnly(numeric_only)
+
         agg_args = tuple() if agg_args is None else agg_args
         agg_kwargs = dict() if agg_kwargs is None else agg_kwargs
 
-        if numeric_only is None:
+        if numeric_only is NumericOnly.AUTO:
             # pandas behavior: if `numeric_only` wasn't explicitly specified then
             # the parameter is considered to be `False` if there are no numeric types
             # in the frame and `True` otherwise.
-            numeric_only = any(
-                is_numeric_dtype(dtype) for dtype in self._query_compiler.dtypes
+            numeric_only = NumericOnly(
+                any(is_numeric_dtype(dtype) for dtype in self._query_compiler.dtypes)
             )
 
-        if numeric_only and self.ndim == 2:
+        if numeric_only > NumericOnly.FALSE and self.ndim == 2:
             by_cols = self._internal_by
             mask_cols = [
                 col
@@ -1064,7 +1063,8 @@ class DataFrameGroupBy(ClassLogger):
                 if (
                     is_numeric_dtype(dtype)
                     or (
-                        isinstance(dtype, pandas.CategoricalDtype)
+                        numeric_only != NumericOnly.TRUE_EXCL_NUMERIC_CATEGORIES
+                        and isinstance(dtype, pandas.CategoricalDtype)
                         and is_numeric_dtype(dtype.categories.dtype)
                     )
                     or col in by_cols