--- a/modin/pandas/groupby.py
+++ b/modin/pandas/groupby.py
@@ -38,7 +38,7 @@ from modin.core.storage_formats.base.query_compiler import BaseQueryCompiler
 from modin.core.dataframe.algebra.default2pandas.groupby import GroupBy
 from modin.config import IsExperimental
 from .series import Series
-from .utils import is_label, NumericOnly
+from .utils import is_label
 
 
 @_inherit_docstrings(pandas.core.groupby.DataFrameGroupBy)
@@ -120,11 +120,38 @@ class DataFrameGroupBy(ClassLogger):
         return len(self)
 
     def skew(self, *args, **kwargs):
-        return self._wrap_aggregation(
-            type(self._query_compiler).groupby_skew,
+        # The 'skew' aggregation is less tolerant to non-numeric columns than others
+        # (i.e. it doesn't allow numeric categoricals), thus dropping non-numeric
+        # columns here since `._wrap_aggregation(numeric_only=True, ...)` is not enough
+        if self.ndim == 2:
+            by_cols = self._internal_by
+            mask_cols = [
+                col
+                for col, dtype in self._df.dtypes.items()
+                if is_numeric_dtype(dtype) or col in by_cols
+            ]
+            if not self._df.columns.equals(mask_cols):
+                masked_df = self._df[mask_cols]
+                masked_obj = type(self)(
+                    df=masked_df,
+                    by=self._by,
+                    axis=self._axis,
+                    idx_name=self._idx_name,
+                    drop=self._drop,
+                    squeeze=self._squeeze,
+                    **self._kwargs,
+                )
+            else:
+                masked_obj = self
+        else:
+            masked_obj = self
+
+        return masked_obj._wrap_aggregation(
+            type(masked_obj._query_compiler).groupby_skew,
             agg_args=args,
             agg_kwargs=kwargs,
-            numeric_only=NumericOnly.TRUE_EXCL_NUMERIC_CATEGORIES,
+            # Don't want to try to drop non-numeric columns for the second time
+            numeric_only=False,
         )
 
     def ffill(self, limit=None):
@@ -1015,7 +1042,7 @@ class DataFrameGroupBy(ClassLogger):
     def _wrap_aggregation(
         self,
         qc_method,
-        numeric_only=NumericOnly.AUTO,
+        numeric_only=None,
         agg_args=None,
         agg_kwargs=None,
         **kwargs,
@@ -1027,8 +1054,12 @@ class DataFrameGroupBy(ClassLogger):
         ----------
         qc_method : callable
             The query compiler method to call.
-        numeric_only : NumericOnly or its values, default: NumericOnly.AUTO
-            Specifies whether to aggregate non numeric columns.
+        numeric_only : {None, True, False}, default: None
+            Specifies whether to aggregate non numeric columns:
+                - True: include only numeric columns (including categories that holds a numeric dtype)
+                - False: include all columns
+                - None: infer the parameter, ``False`` if there are no numeric types in the frame,
+                  ``True`` otherwise.
         agg_args : list-like, optional
             Positional arguments to pass to the aggregation function.
         agg_kwargs : dict-like, optional
@@ -1041,20 +1072,18 @@ class DataFrameGroupBy(ClassLogger):
         DataFrame or Series
             Returns the same type as `self._df`.
         """
-        numeric_only = NumericOnly(numeric_only)
-
         agg_args = tuple() if agg_args is None else agg_args
         agg_kwargs = dict() if agg_kwargs is None else agg_kwargs
 
-        if numeric_only is NumericOnly.AUTO:
+        if numeric_only is None:
             # pandas behavior: if `numeric_only` wasn't explicitly specified then
             # the parameter is considered to be `False` if there are no numeric types
             # in the frame and `True` otherwise.
-            numeric_only = NumericOnly(
-                any(is_numeric_dtype(dtype) for dtype in self._query_compiler.dtypes)
+            numeric_only = any(
+                is_numeric_dtype(dtype) for dtype in self._query_compiler.dtypes
             )
 
-        if numeric_only > NumericOnly.FALSE and self.ndim == 2:
+        if numeric_only and self.ndim == 2:
             by_cols = self._internal_by
             mask_cols = [
                 col
@@ -1062,8 +1091,7 @@ class DataFrameGroupBy(ClassLogger):
                 if (
                     is_numeric_dtype(dtype)
                     or (
-                        numeric_only != NumericOnly.TRUE_EXCL_NUMERIC_CATEGORIES
-                        and isinstance(dtype, pandas.CategoricalDtype)
+                        isinstance(dtype, pandas.CategoricalDtype)
                         and is_numeric_dtype(dtype.categories.dtype)
                     )
                     or col in by_cols