@dchigarev 2022-12-08
@vnlitvinov 

> I wonder if it's possible to use old implementation for mostly square dataframes to not introduce such a speed loss...

I've created a separate issue (#5394) as this perf-issue affects every Map/MapReduce function so there should be a general solution for this.